FIX: Test for being a react component was wrong
It's a bit tricky to test for that, and documentation for TestUtils (https://facebook.github.io/react/docs/test-utils.html) is lacking examples (as of the time of this commit).

By comparing it with Virtual DOM glossary  (https://facebook.github.io/react/docs/glossary.html#react-components) one can assume that instantiating it with new keyword and then checking with TestUtils.isCompositeComponent is a way to go.

That's how I do it ATM.